Convert a chunk to a string.
In latest node versions it's possible that chunk is an object when using this module with gulp streams as vinyl files return objects. Because of this the module breaks, as a sanity check we should be converting the chunk to a string to ensure a buffer can be created from it.